requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
base58==2.1.1
bitarray==2.9.2
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
cytoolz==0.12.3
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.10
ipfshttpclient==0.7.0
multiaddr==0.0.9
multidict==6.1.0
mysql-connector-python==9.0.0
netaddr==1.3.0
packaging==23.2
parsimonious==0.10.0
py-solc-x==2.0.3
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyunormalize==16.0.0
pywin32==306
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
six==1.16.0
toolz==0.12.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
varint==1.0.2
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
